Move "Metric" and "Mega" into a separate source file
